+++ resources/js/Components/Buttons/PrimaryButton.vue
