+++ resources/js/Components/Cards/Welcome.vue
@@ -1,5 +1,5 @@
 <script setup>
-    import ApplicationLogo from '@/Components/ApplicationLogo.vue';
+    import ApplicationLogo from '@/Components/Utils/ApplicationLogo.vue';
 </script>
 
 <template>
