+++ resources/js/Components/Modals/ConfirmationModal.vue
