+++ resources/js/Components/Modals/ConfirmsPassword.vue
@@ -1,6 +1,6 @@
 <script setup>
     import { ref, reactive, nextTick } from 'vue';
-    import DialogModal from './DialogModal.vue';
+    import DialogModal from './Modals/DialogModal.vue';
     import InputError from './InputError.vue';
     import PrimaryButton from './PrimaryButton.vue';
     import SecondaryButton from './SecondaryButton.vue';
