+++ resources/js/Components/Sections/SectionTitle.vue
