+++ resources/js/Components/Utils/ApplicationLogo.vue
