+++ resources/js/Components/Utils/ApplicationMark.vue
