+++ resources/js/Components/Utils/AuthenticationCardLogo.vue
