+++ resources/js/Components/Utils/Banner.vue
