+++ resources/js/Layouts/AppLayout.vue
@@ -1,12 +1,12 @@
 <script setup>
     import { ref } from 'vue';
     import { Head, Link, router } from '@inertiajs/vue3';
-    import ApplicationMark from '@/Components/ApplicationMark.vue';
-    import Banner from '@/Components/Banner.vue';
-    import Dropdown from '@/Components/Dropdown.vue';
-    import DropdownLink from '@/Components/DropdownLink.vue';
-    import NavLink from '@/Components/NavLink.vue';
-    import ResponsiveNavLink from '@/Components/ResponsiveNavLink.vue';
+    import ApplicationMark from '@/Components/Utils/ApplicationMark.vue';
+    import Banner from '@/Components/Utils/Banner.vue';
+    import Dropdown from '@/Components/Utils/Dropdown.vue';
+    import DropdownLink from '@/Components/Utils/DropdownLink.vue';
+    import NavLink from '@/Components/Utils/NavLink.vue';
+    import ResponsiveNavLink from '@/Components/Utils/ResponsiveNavLink.vue';
 
     defineProps({
         title: String,
