+++ resources/js/Pages/API/Partials/ApiTokenManager.vue
@@ -1,19 +1,19 @@
 <script setup>
     import { ref } from 'vue';
     import { useForm } from '@inertiajs/vue3';
-    import ActionMessage from '@/Components/ActionMessage.vue';
-    import ActionSection from '@/Components/ActionSection.vue';
-    import Checkbox from '@/Components/Checkbox.vue';
-    import ConfirmationModal from '@/Components/ConfirmationModal.vue';
-    import DangerButton from '@/Components/DangerButton.vue';
-    import DialogModal from '@/Components/DialogModal.vue';
+    import ActionMessage from '@/Components/Actions/ActionMessage.vue';
+    import ActionSection from '@/Components/Actions/ActionSection.vue';
+    import Checkbox from '@/Components/Inputs/Checkbox.vue';
+    import ConfirmationModal from '@/Components/Modals/ConfirmationModal.vue';
+    import DangerButton from '@/Components/Buttons/DangerButton.vue';
+    import DialogModal from '@/Components/Modals/DialogModal.vue';
     import FormSection from '@/Components/FormSection.vue';
-    import InputError from '@/Components/InputError.vue';
-    import InputLabel from '@/Components/InputLabel.vue';
-    import PrimaryButton from '@/Components/PrimaryButton.vue';
-    import SecondaryButton from '@/Components/SecondaryButton.vue';
+    import InputError from '@/Components/Inputs/InputError.vue';
+    import InputLabel from '@/Components/Inputs/InputLabel.vue';
+    import PrimaryButton from '@/Components/Buttons/PrimaryButton.vue';
+    import SecondaryButton from '@/Components/Buttons/SecondaryButton.vue';
     import SectionBorder from '@/Components/SectionBorder.vue';
-    import TextInput from '@/Components/TextInput.vue';
+    import TextInput from '@/Components/Inputs/TextInput.vue';
 
     const props = defineProps({
         tokens: Array,
