+++ resources/js/Pages/Auth/ForgotPassword.vue
@@ -1,11 +1,11 @@
 <script setup>
     import { Head, useForm } from '@inertiajs/vue3';
-    import AuthenticationCard from '@/Components/AuthenticationCard.vue';
-    import AuthenticationCardLogo from '@/Components/AuthenticationCardLogo.vue';
-    import InputError from '@/Components/InputError.vue';
-    import InputLabel from '@/Components/InputLabel.vue';
-    import PrimaryButton from '@/Components/PrimaryButton.vue';
-    import TextInput from '@/Components/TextInput.vue';
+    import AuthenticationCard from '@/Components/Cards/AuthenticationCard.vue';
+    import AuthenticationCardLogo from '@/Components/Utils/AuthenticationCardLogo.vue';
+    import InputError from '@/Components/Inputs/InputError.vue';
+    import InputLabel from '@/Components/Inputs/InputLabel.vue';
+    import PrimaryButton from '@/Components/Buttons/PrimaryButton.vue';
+    import TextInput from '@/Components/Inputs/TextInput.vue';
 
     defineProps({
         status: String,
