+++ resources/js/Pages/Profile/Partials/DeleteUserForm.vue
@@ -1,12 +1,12 @@
 <script setup>
     import { ref } from 'vue';
     import { useForm } from '@inertiajs/vue3';
-    import ActionSection from '@/Components/ActionSection.vue';
-    import DangerButton from '@/Components/DangerButton.vue';
-    import DialogModal from '@/Components/DialogModal.vue';
-    import InputError from '@/Components/InputError.vue';
-    import SecondaryButton from '@/Components/SecondaryButton.vue';
-    import TextInput from '@/Components/TextInput.vue';
+    import ActionSection from '@/Components/Actions/ActionSection.vue';
+    import DangerButton from '@/Components/Buttons/DangerButton.vue';
+    import DialogModal from '@/Components/Modals/DialogModal.vue';
+    import InputError from '@/Components/Inputs/InputError.vue';
+    import SecondaryButton from '@/Components/Buttons/SecondaryButton.vue';
+    import TextInput from '@/Components/Inputs/TextInput.vue';
 
     const confirmingUserDeletion = ref(false);
     const passwordInput = ref(null);
