+++ resources/js/Pages/Profile/Partials/UpdatePasswordForm.vue
@@ -1,12 +1,12 @@
 <script setup>
     import { ref } from 'vue';
     import { useForm } from '@inertiajs/vue3';
-    import ActionMessage from '@/Components/ActionMessage.vue';
+    import ActionMessage from '@/Components/Actions/ActionMessage.vue';
     import FormSection from '@/Components/FormSection.vue';
-    import InputError from '@/Components/InputError.vue';
-    import InputLabel from '@/Components/InputLabel.vue';
-    import PrimaryButton from '@/Components/PrimaryButton.vue';
-    import TextInput from '@/Components/TextInput.vue';
+    import InputError from '@/Components/Inputs/InputError.vue';
+    import InputLabel from '@/Components/Inputs/InputLabel.vue';
+    import PrimaryButton from '@/Components/Buttons/PrimaryButton.vue';
+    import TextInput from '@/Components/Inputs/TextInput.vue';
 
     const passwordInput = ref(null);
     const currentPasswordInput = ref(null);
