+++ resources/js/Pages/TermsOfService.vue
@@ -1,6 +1,6 @@
 <script setup>
     import { Head } from '@inertiajs/vue3';
-    import AuthenticationCardLogo from '@/Components/AuthenticationCardLogo.vue';
+    import AuthenticationCardLogo from '@/Components/Utils/AuthenticationCardLogo.vue';
 
     defineProps({
         terms: String,
